% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_matrix.R
\name{corr_matrix}
\alias{corr_matrix}
\title{Generation of a correlation matrix}
\usage{
corr_matrix(exp, r)
}
\arguments{
\item{exp}{Expression matrix with genes as rows and samples as columns [data.frame]}

\item{r}{Correlation cut-off to use. Correlations below the defined cut-off will be replaced with zero before returning the final correlation matrix.}
}
\value{
Gene-gene correlation matrix of the ImSig genes.
}
\description{
Creates a correlation matrix using ImSig signature genes.
}
