% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_cli.R
\docType{data}
\name{example_cli}
\alias{example_cli}
\title{Example clinical data file for survival analysis with ImSig}
\format{dataframe}
\usage{
example_cli
}
\description{
An example clinical data file. Minimum required informations are the sample name (same as that of the expression matrix), event (dead or alive) and time to event (days, months or years).
}
\keyword{datasets}
